+++ src/main/java/org/truffleruby/core/encoding/Encodings.java
@@ -171,4 +171,5 @@ public final class Encodings {
 
         return null;
     }
+
 }