+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Compatibility:
 
 Performance:
 
+* Optimize encoding negotiation for ASCII-compatible encodings (@eregon, @andrykonchin).
+
 Changes:
 * Inherit `Polyglot::ForeignException` from `StandardError` instead of `Exception` (#3620, @andrykonchin).